Updates priority values
